+++ src/ViewInstance.js
@@ -1,14 +1,8 @@
 import {
-  concat,
-  filter,
   get,
   has,
   cloneDeep,
   orderBy,
-  map,
-  omit,
-  reject,
-  set,
   isEmpty,
   values,
 } from 'lodash';
@@ -57,12 +51,12 @@ import { ViewMetaData } from '@folio/stripes/smart-components';
 import {
   areAllFieldsEmpty,
   craftLayerUrl,
-  psTitleRelationshipId,
   checkIfElementIsEmpty,
   convertArrayToBlocks,
   checkIfArrayIsEmpty,
   staffOnlyFormatter,
   getSortedNotes,
+  parseTitles,
 } from './utils';
 import formatters from './referenceFormatters';
 import Holdings from './Holdings';
@@ -72,6 +66,8 @@ import ViewHoldingsRecord from './ViewHoldingsRecord';
 import ViewMarc from './ViewMarc';
 import makeConnectedInstance from './ConnectedInstance';
 import withLocation from './withLocation';
+
+import { TitlesView } from './components';
 import {
   wrappingCell,
   noValue,
@@ -191,48 +187,14 @@ class ViewInstance extends React.Component {
   };
 
   update = (instance) => {
-    this.combineRelTitles(instance);
+    // Massage record to add preceeding and succeeding title fields
+    parseTitles(instance, 'precedingTitles', 'precedingInstanceId');
+    parseTitles(instance, 'succeedingTitles', 'succeedingInstanceId');
+
     this.props.mutator.selectedInstance.PUT(instance).then(() => {
       this.resetLayerQueryParam();
     });
   };
-
-  // Combine precedingTitles with parentInstances, and succeedingTitles with childInstances,
-  // before saving the instance record
-  combineRelTitles = (instance) => {
-    // preceding/succeeding titles are stored in parentInstances and childInstances
-    // in the instance record and needfrle to be combined with existing relationships here
-    // before saving. Each title needs to provide an instance relationship
-    // type ID corresponding to 'preceeding-succeeding' in addition to the actual parent
-    // instance ID.
-    let instanceCopy = instance;
-    const titleRelationshipTypeId = psTitleRelationshipId(this.props.resources.instanceRelationshipTypes.records);
-    const precedingTitles = map(instanceCopy.precedingTitles, p => { p.instanceRelationshipTypeId = titleRelationshipTypeId; return p; });
-    const succeedingTitles = map(instanceCopy.succeedingTitles, p => { p.instanceRelationshipTypeId = titleRelationshipTypeId; return p; });
-    set(instanceCopy, 'parentInstances', concat(instanceCopy.parentInstances, precedingTitles));
-    set(instanceCopy, 'childInstances', concat(instanceCopy.childInstances, succeedingTitles));
-    instanceCopy = omit(instanceCopy, ['precedingTitles', 'succeedingTitles']);
-    return instanceCopy;
-  };
-
-  // Separate preceding/succeeding title relationships from other types of
-  // parent/child instances before displaying the record
-  splitRelTitles = (instance) => {
-    const instanceCopy = cloneDeep(instance);
-    const psRelId = psTitleRelationshipId(this.props.resources.instanceRelationshipTypes.records);
-    if (instanceCopy.parentInstances) {
-      const parentInstances = reject(instanceCopy.parentInstances, { 'instanceRelationshipTypeId': psRelId });
-      const precedingTitles = filter(instanceCopy.parentInstances, { 'instanceRelationshipTypeId': psRelId });
-      instance.precedingTitles = instance.precedingTitles || precedingTitles;
-      instance.parentInstances = parentInstances;
-    }
-    if (instanceCopy.childInstances) {
-      const childInstances = reject(instanceCopy.childInstances, { 'instanceRelationshipTypeId': psRelId });
-      const succeedingTitles = filter(instanceCopy.childInstances, { 'instanceRelationshipTypeId': psRelId });
-      instance.succeedingTitles = instance.succeedingTitles || succeedingTitles;
-      instance.childInstances = childInstances;
-    }
-  }
 
   resetLayerQueryParam = (e) => {
     if (e) e.preventDefault();
@@ -262,7 +224,6 @@ class ViewInstance extends React.Component {
     this.resetLayerQueryParam();
     this.props.goTo(`/inventory/view/${this.props.match.params.id}${search}`);
   };
-
 
   createHoldingsRecord = (holdingsRecord) => {
     // POST holdings record
@@ -561,8 +522,6 @@ class ViewInstance extends React.Component {
       </FormattedMessage>
     );
 
-    this.splitRelTitles(instance);
-
     if (query.layer === 'edit') {
       return (
         <IntlConsumer>
@@ -667,16 +626,13 @@ class ViewInstance extends React.Component {
       alternativeTitles: get(instance, ['alternativeTitles'], []),
       indexTitle: get(instance, ['indexTitle'], '-'),
       series: get(instance, ['series'], []),
-      precedingTitles: get(instance, ['precedingTitles'], []),
-      succeedingTitles: get(instance, ['succeedingTitles'], []),
     };
+    const {
+      precedingTitles,
+      succeedingTitles,
+    } = instance;
 
     const seriesContent = !isEmpty(titleData.series) ? titleData.series.map(x => ({ value: x })) : emptyList;
-
-    const precedingTitlesContent = !isEmpty(titleData.precedingTitles) ? formatters.precedingTitlesFormatter(instance, location) : noValue;
-
-    const succeedingTitlesContent = !isEmpty(titleData.succeedingTitles) ? formatters.succeedingTitlesFormatter(instance, location) : noValue;
-
     const identifiers = get(instance, 'identifiers', []).map(x => ({
       identifierType: this.refLookup(referenceTables.identifierTypes, get(x, 'identifierTypeId')).name,
       value: x.value,
@@ -1080,9 +1036,11 @@ class ViewInstance extends React.Component {
               data-test-preceding-titles
               xs={12}
             >
-              <KeyValue
+              <TitlesView
+                id="precedingTitles"
+                titleKey="precedingInstanceId"
                 label={<FormattedMessage id="ui-inventory.precedingTitles" />}
-                value={precedingTitlesContent}
+                titles={isEmpty(precedingTitles) ? emptyList : precedingTitles}
               />
             </Col>
           </Row>
@@ -1091,9 +1049,11 @@ class ViewInstance extends React.Component {
               data-test-succeeding-titles
               xs={12}
             >
-              <KeyValue
+              <TitlesView
+                id="succeedingTitles"
+                titleKey="succeedingInstanceId"
                 label={<FormattedMessage id="ui-inventory.succeedingTitles" />}
-                value={succeedingTitlesContent}
+                titles={isEmpty(succeedingTitles) ? emptyList : succeedingTitles}
               />
             </Col>
           </Row>