+++ src/components/InstancesList/InstancesList.js
@@ -15,7 +15,10 @@ import {
 
 import { AppIcon } from '@folio/stripes/core';
 import { SearchAndSort } from '@folio/stripes/smart-components';
-import { Button } from '@folio/stripes/components';
+import {
+  Button,
+  Icon
+} from '@folio/stripes/components';
 
 import FilterNavigation from '../FilterNavigation';
 import packageInfo from '../../../package';
@@ -32,6 +35,8 @@ import {
   InstancesIdReport,
 } from '../../reports';
 import ErrorModal from '../ErrorModal';
+
+import css from './instances.css';
 
 const INITIAL_RESULT_COUNT = 30;
 const RESULT_COUNT_INCREMENT = 30;
@@ -181,6 +186,11 @@ class InstancesView extends React.Component {
             this.startInTransitReportGeneration();
           }}
         >
+          <Icon
+            icon="reports"
+            size="medium"
+            iconClassName={css.actionIcon}
+          />
           <FormattedMessage id="ui-inventory.inTransitReport" />
         </Button>
         <Button
@@ -193,6 +203,11 @@ class InstancesView extends React.Component {
             this.generateInstancesIdReport();
           }}
         >
+          <Icon
+            icon="save"
+            size="medium"
+            iconClassName={css.actionIcon}
+          />
           <FormattedMessage id="ui-inventory.saveInstancesUIIDS" />
         </Button>
       </Fragment>