+++ src/components/index.js
@@ -4,4 +4,9 @@ export { default as InstanceFilters, instanceFilterRenderer } from './InstanceFi
 export { default as HoldingsRecordFilters, holdingsRecordFilterRenderer } from './HoldingsRecordFilters';
 export { default as ItemFilters, itemFilterRenderer } from './ItemFilters';
 export { default as InstancesList } from './InstancesList';
+export { default as InstancePlugin } from './InstancePlugin';
+export { default as ConnectedTitle } from './ConnectedTitle';
+export { default as UnconnectedTitle } from './UnconnectedTitle';
+export { default as TitleLabel } from './TitleLabel';
+export { default as TitlesView } from './TitlesView';
 export { PaneLoading, ViewLoading } from './Loading';