+++ test/bigtest/tests/instance-edit-page-test.js
@@ -1,12 +1,31 @@
+import React from 'react';
 import { beforeEach, describe, it } from '@bigtest/mocha';
 import { expect } from 'chai';
+
+import { Button } from '@folio/stripes/components';
 
 import setupApplication from '../helpers/setup-application';
 import InstanceEditPage from '../interactors/instance-edit-page';
 import InstanceViewPage from '../interactors/instance-view-page';
 
 describe('InstanceEditPage', () => {
-  setupApplication();
+  setupApplication({
+    modules: [{
+      type: 'plugin',
+      name: '@folio/plugin-find-instance',
+      displayName: 'Find instance',
+      pluginType: 'find-instance',
+      /* eslint-disable-next-line react/prop-types */
+      module: ({ selectInstance }) => (
+        <Button
+          data-test-plugin-find-record-button
+          onClick={() => selectInstance({ id: 1, title: 'Fake instance' })}
+        >
+          +
+        </Button>
+      ),
+    }],
+  });
 
   beforeEach(async function () {
     const instance = this.server.create('instance');
@@ -157,6 +176,28 @@ describe('InstanceEditPage', () => {
 
       it('should change the button style to "primary"', () => {
         expect(InstanceEditPage.contributors.firstContributorIsPrimary).to.be.true;
+      });
+    });
+  });
+
+  describe('clicking on "add preceding title"', () => {
+    const prevCount = InstanceEditPage.precedingTitles.precedingTitlesCount;
+
+    beforeEach(async () => {
+      await InstanceEditPage.precedingTitles.clickAddPrecedingTitle();
+    });
+
+    it('should increase number of preceding title"', () => {
+      expect(InstanceEditPage.precedingTitles.precedingTitlesCount).to.be.gt(prevCount);
+    });
+
+    describe('clicking add instance', () => {
+      beforeEach(async () => {
+        await InstanceEditPage.precedingTitles.clickAddInstance();
+      });
+
+      it('should add instance', () => {
+        expect(InstanceEditPage.precedingTitles.instanceName).to.be.equal('Fake instance');
       });
     });
   });